+++ torchy_baselines/common/vec_env/__init__.py
@@ -3,8 +3,8 @@ import typing
 from typing import Optional
 from copy import deepcopy
 
-from torchy_baselines.common.vec_env.base_vec_env import AlreadySteppingError, NotSteppingError,\
-    VecEnv, VecEnvWrapper, CloudpickleWrapper
+from torchy_baselines.common.vec_env.base_vec_env import (AlreadySteppingError, NotSteppingError,
+                                                          VecEnv, VecEnvWrapper, CloudpickleWrapper)
 from torchy_baselines.common.vec_env.dummy_vec_env import DummyVecEnv
 from torchy_baselines.common.vec_env.subproc_vec_env import SubprocVecEnv
 from torchy_baselines.common.vec_env.vec_frame_stack import VecFrameStack